+++ packages/backend-defaults/src/entrypoints/urlReader/lib/GithubUrlReader.test.ts
@@ -492,7 +492,7 @@ describe('GithubUrlReader', () => {
     });
 
     it('should override the token when provided', async () => {
-      expect.assertions(1);
+      expect.assertions(2);
 
       const mockHeaders = {
         Authorization: 'bearer blah',
@@ -503,6 +503,20 @@ describe('GithubUrlReader', () => {
       });
 
       worker.use(
+        rest.get(
+          'https://ghe.github.com/api/v3/repos/backstage/mock/commits/main/status',
+          (req, res, ctx) => {
+            expect(req.headers.get('authorization')).toBe(
+              'Bearer overridentoken',
+            );
+
+            return res(
+              ctx.status(200),
+              ctx.set('Content-Type', 'application/json'),
+              ctx.json(commitStatusGheResponse),
+            );
+          },
+        ),
         rest.get(
           'https://ghe.github.com/api/v3/repos/backstage/mock/tarball/etag123abc',
           (req, res, ctx) => {
@@ -678,6 +692,21 @@ describe('GithubUrlReader', () => {
       },
     ];
 
+    const gheCommitsResponse = {
+      sha: 'etag123abc',
+      repository: {
+        id: 123,
+        full_name: 'backstage/mock',
+        default_branch: 'main',
+        branches_url:
+          'https://ghe.github.com/api/v3/repos/backstage/mock/branches{/branch}',
+        archive_url:
+          'https://ghe.github.com/api/v3/repos/backstage/mock/{archive_format}{/ref}',
+        trees_url:
+          'https://ghe.github.com/api/v3/repos/backstage/mock/git/trees{/sha}',
+      },
+    } as Partial<GhCombinedCommitStatusResponse>;
+
     // Tarballs
     beforeEach(() => {
       worker.use(
@@ -773,21 +802,6 @@ describe('GithubUrlReader', () => {
         },
       } as Partial<GhCombinedCommitStatusResponse>;
 
-      const gheResponse = {
-        sha: 'etag123abc',
-        repository: {
-          id: 123,
-          full_name: 'backstage/mock',
-          default_branch: 'main',
-          branches_url:
-            'https://ghe.github.com/api/v3/repos/backstage/mock/branches{/branch}',
-          archive_url:
-            'https://ghe.github.com/api/v3/repos/backstage/mock/{archive_format}{/ref}',
-          trees_url:
-            'https://ghe.github.com/api/v3/repos/backstage/mock/git/trees{/sha}',
-        },
-      } as Partial<GhCombinedCommitStatusResponse>;
-
       worker.use(
         rest.get(
           'https://api.github.com/repos/backstage/mock/commits/main/status',
@@ -810,7 +824,7 @@ describe('GithubUrlReader', () => {
               return res(
                 ctx.status(200),
                 ctx.set('Content-Type', 'application/json'),
-                ctx.json(gheResponse),
+                ctx.json(gheCommitsResponse),
               );
             }
 
@@ -969,9 +983,23 @@ describe('GithubUrlReader', () => {
     });
 
     it('passes through a token for the search request', async () => {
-      expect.assertions(1);
+      expect.assertions(2);
 
       worker.use(
+        rest.get(
+          'https://ghe.github.com/api/v3/repos/backstage/mock/commits/main/status',
+          (req, res, ctx) => {
+            expect(req.headers.get('authorization')).toBe(
+              'Bearer overridentoken',
+            );
+
+            return res(
+              ctx.status(200),
+              ctx.set('Content-Type', 'application/json'),
+              ctx.json(gheCommitsResponse),
+            );
+          },
+        ),
         rest.get(
           'https://ghe.github.com/api/v3/repos/backstage/mock/git/trees/etag123abc',
           (req, res, ctx) => {