+++ packages/frontend-app-api/src/wiring/createSpecializedApp.tsx
@@ -195,7 +195,7 @@ class RouteResolutionApiProxy implements RouteResolutionApi {
 
 /**
  * Creates an empty app without any default features. This is a low-level API is
- * intended for use in tests or specialized setups. Typically wou want to use
+ * intended for use in tests or specialized setups. Typically you want to use
  * `createApp` from `@backstage/frontend-defaults` instead.
  *
  * @public