+++ plugins/events-backend-module-github/src/index.ts
@@ -20,6 +20,7 @@
  *
  * @packageDocumentation
  */
+
 import { createBackendFeatureLoader } from '@backstage/backend-plugin-api';
 
 export default createBackendFeatureLoader({
@@ -31,5 +32,8 @@ export default createBackendFeatureLoader({
   },
 });
 
-export { createGithubSignatureValidator } from './http/createGithubSignatureValidator';
+// TODO(freben): This is not exported at the moment since it depends on the octokit provider.
+// Until we have made that a core thing in integrations, we can't export it
+// export { createGithubSignatureValidator } from './http/createGithubSignatureValidator';
+
 export { GithubEventRouter } from './router/GithubEventRouter';