+++ plugins/org/src/components/MyGroupsSidebarItem/MyGroupsSidebarItem.tsx
@@ -66,7 +66,7 @@ export const MyGroupsSidebarItem = (props: {
           ...(filter ?? {}),
         },
       ],
-      fields: ['metadata', 'kind'],
+      fields: ['metadata', 'kind', 'spec.profile'],
     });
     return response.items;
   }, []);