+++ plugins/scaffolder/src/components/fields/RepoBranchPicker/GitHubRepoBranchPicker.test.tsx
@@ -113,7 +113,7 @@ describe('GitHubRepoBranchPicker', () => {
       </TestApiProvider>,
     );
 
-    // Open the Autcomplete dropdown
+    // Open the Autocomplete dropdown
     const input = getByRole('textbox');
     await userEvent.click(input);