+++ plugins/app-visualizer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-app-visualizer
 
+## 0.1.19-next.0
+
+### Patch Changes
+
+- fb58f20: Internal update to use the new `pluginId` option of `createFrontendPlugin`.
+- Updated dependencies
+  - @backstage/core-components@0.17.2-next.0
+  - @backstage/frontend-plugin-api@0.10.2-next.0
+  - @backstage/core-plugin-api@1.10.6
+
 ## 0.1.18
 
 ### Patch Changes