+++ plugins/events-backend-module-google-pubsub/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-events-backend-module-google-pubsub
 
+## 0.1.0-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.3.1-next.1
+  - @backstage/config@1.3.2
+  - @backstage/errors@1.2.7
+  - @backstage/types@1.2.1
+  - @backstage/plugin-events-node@0.4.11-next.1
+
 ## 0.1.0-next.0
 
 ### Minor Changes